Module google_customsearch1::api

Structs

CseListCall
Returns metadata about the search performed, metadata about the engine used for the search, and the search results.

CseMethods
A builder providing access to all methods supported on cse resources. It is not used directly, but through the CustomSearchAPI hub.

CseSiterestrictListCall
Returns metadata about the search performed, metadata about the engine used for the search, and the search results. Uses a small set of url patterns.

CustomSearchAPI
Central instance to access all CustomSearchAPI related resource activities

Promotion
Promotion result.

PromotionBodyLines
Block object belonging to a promotion.

PromotionImage
Image belonging to a promotion.

Result
A custom search result.

ResultImage
Image belonging to a custom search result.

ResultLabels
Refinement label associated with a custom search result.

Search
Response to a custom search request.

SearchQueries
Query metadata for the previous, current, and next pages of results.

SearchQueriesNextPage
Custom search request metadata.

SearchQueriesPreviousPage
Custom search request metadata.

SearchQueriesRequest
Custom search request metadata.

SearchSearchInformation
Metadata about a search operation.

SearchSpelling
Spell correction information for a query.

SearchUrl
OpenSearch template and URL.