Struct google_customsearch1::api::SearchQueriesNextPage

pub struct SearchQueriesNextPage {
    pub count: Option<i32>,
    pub cr: Option<String>,
    pub cx: Option<String>,
    pub date_restrict: Option<String>,
    pub disable_cn_tw_translation: Option<String>,
    pub exact_terms: Option<String>,
    pub exclude_terms: Option<String>,
    pub file_type: Option<String>,
    pub filter: Option<String>,
    pub gl: Option<String>,
    pub google_host: Option<String>,
    pub high_range: Option<String>,
    pub hl: Option<String>,
    pub hq: Option<String>,
    pub img_color_type: Option<String>,
    pub img_dominant_color: Option<String>,
    pub img_size: Option<String>,
    pub img_type: Option<String>,
    pub input_encoding: Option<String>,
    pub language: Option<String>,
    pub link_site: Option<String>,
    pub low_range: Option<String>,
    pub or_terms: Option<String>,
    pub output_encoding: Option<String>,
    pub related_site: Option<String>,
    pub rights: Option<String>,
    pub safe: Option<String>,
    pub search_terms: Option<String>,
    pub search_type: Option<String>,
    pub site_search: Option<String>,
    pub site_search_filter: Option<String>,
    pub sort: Option<String>,
    pub start_index: Option<i32>,
    pub start_page: Option<i32>,
    pub title: Option<String>,
    pub total_results: Option<String>,
}

Custom search request metadata.

This type is not used in any activity, and only used as part of another schema.

Fields
count: Option<i32>

Number of search results returned in this set.

cr: Option<String>

Restricts search results to documents originating in a particular country. You may use Boolean operators in the cr parameter's value. Google WebSearch determines the country of a document by analyzing the following: * The top-level domain (TLD) of the document's URL. * The geographic location of the web server's IP address. See Country (cr) Parameter Values for a list of valid values for this parameter.

cx: Option<String>

The identifier of an engine created using the Programmable Search Engine Control Panel. This is a custom property not defined in the OpenSearch spec. This parameter is required.

date_restrict: Option<String>

Restricts results to URLs based on date. Supported values include: * d[number]: requests results from the specified number of past days. * w[number]: requests results from the specified number of past weeks. * m[number]: requests results from the specified number of past months. * y[number]: requests results from the specified number of past years.

disable_cn_tw_translation: Option<String>

Enables or disables the Simplified and Traditional Chinese Search feature. Supported values are: * 0: enabled (default) * 1: disabled

exact_terms: Option<String>

Identifies a phrase that all documents in the search results must contain.

exclude_terms: Option<String>

Identifies a word or phrase that should not appear in any documents in the search results.

file_type: Option<String>

Restricts results to files of a specified extension. Filetypes supported by Google include: * Adobe Portable Document Format (pdf) * Adobe PostScript (ps) * Lotus 1-2-3 (wk1, wk2, wk3, wk4, wk5, wki, wks, wku) * Lotus WordPro (lwp) * Macwrite (mw) * Microsoft Excel (xls) * Microsoft PowerPoint (ppt) * Microsoft Word (doc) * Microsoft Works (wks, wps, wdb) * Microsoft Write (wri) * Rich Text Format (rtf) * Shockwave Flash (swf) * Text (ans, txt). Additional filetypes may be added in the future. An up-to-date list can always be found in Google's file type FAQ.

filter: Option<String>

Activates or deactivates the automatic filtering of Google search results. See Automatic Filtering for more information about Google's search results filters. Valid values for this parameter are: * 0: Disabled * 1: Enabled (default) Note: By default, Google applies filtering to all search results to improve the quality of those results.

gl: Option<String>

Boosts search results whose country of origin matches the parameter value. See Country Codes for a list of valid values. Specifying a gl parameter value in WebSearch requests should improve the relevance of results. This is particularly true for international customers and, even more specifically, for customers in English-speaking countries other than the United States.

google_host: Option<String>

Specifies the Google domain (for example, google.com, google.de, or google.fr) to which the search should be limited.

high_range: Option<String>

Specifies the ending value for a search range. Use cse:lowRange and cse:highrange to append an inclusive search range of lowRange...highRange to the query.

hl: Option<String>

Specifies the interface language (host language) of your user interface. Explicitly setting this parameter improves the performance and the quality of your search results. See the Interface Languages section of Internationalizing Queries and Results Presentation for more information, and Supported Interface Languages for a list of supported languages.

hq: Option<String>

Appends the specified query terms to the query, as if they were combined with a logical AND operator.

img_color_type: Option<String>

Restricts results to images of a specified color type. Supported values are: * mono (black and white) * gray (grayscale) * color (color)

img_dominant_color: Option<String>

Restricts results to images with a specific dominant color. Supported values are: * red * orange * yellow * green * teal * blue * purple * pink * white * gray * black * brown

img_size: Option<String>

Restricts results to images of a specified size. Supported values are: * icon (small) * small | medium | large | xlarge (medium) * xxlarge (large) * huge (extra-large)

img_type: Option<String>

Restricts results to images of a specified type. Supported values are: * clipart (Clip art) * face (Face) * lineart (Line drawing) * photo (Photo) * animated (Animated) * stock (Stock)

input_encoding: Option<String>

The character encoding supported for search requests.

language: Option<String>

The language of the search results.

link_site: Option<String>

Specifies that all results should contain a link to a specific URL.

low_range: Option<String>

Specifies the starting value for a search range. Use cse:lowRange and cse:highrange to append an inclusive search range of lowRange...highRange to the query.

or_terms: Option<String>

Provides additional search terms to check for in a document, where each document in the search results must contain at least one of the additional search terms. You can also use the Boolean OR query term for this type of query.

output_encoding: Option<String>

The character encoding supported for search results.

related_site: Option<String>

Specifies that all search results should be pages that are related to the specified URL. The parameter value should be a URL.

rights: Option<String>

Filters based on licensing. Supported values include: * cc_publicdomain * cc_attribute * cc_sharealike * cc_noncommercial * cc_nonderived

safe: Option<String>

Specifies the SafeSearch level used for filtering out adult results. This is a custom property not defined in the OpenSearch spec. Valid parameter values are: * "off": Disable SafeSearch * "active": Enable SafeSearch

search_terms: Option<String>

The search terms entered by the user.

search_type: Option<String>

Allowed values are web or image. If unspecified, results are limited to webpages.

site_search: Option<String>

Restricts results to URLs from a specified site.

site_search_filter: Option<String>

Specifies whether to include or exclude results from the site named in the sitesearch parameter. Supported values are: * i: include content from site * e: exclude content from site

sort: Option<String>

Specifies that results should be sorted according to the specified expression. For example, sort by date.

start_index: Option<i32>

The index of the current set of search results into the total set of results, where the index of the first result is 1.

start_page: Option<i32>

The page number of this set of results, where the page length is set by the count property.

title: Option<String>

A description of the query.

total_results: Option<String>

Estimated number of total search results. May not be accurate.