Struct google_customsearch1::CseListCall

pub struct CseListCall<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

Returns metadata about the search performed, metadata about the custom search engine used for the search, and the search results.

A builder for the list method supported by a cse resource. It is not used directly, but through a CseMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.cse().list("q")
             .start(88)
             .sort("diam")
             .site_search_filter("justo")
             .site_search("est")
             .search_type("clita")
             .safe("invidunt")
             .rights("ut")
             .related_site("dolores")
             .or_terms("eos")
             .num(23)
             .lr("duo")
             .low_range("sed")
             .link_site("aliquyam")
             .img_type("ea")
             .img_size("ea")
             .img_dominant_color("et")
             .img_color_type("dolor")
             .hq("diam")
             .hl("kasd")
             .high_range("invidunt")
             .googlehost("rebum.")
             .gl("Lorem")
             .filter("clita")
             .file_type("invidunt")
             .exclude_terms("eirmod")
             .exact_terms("At")
             .date_restrict("consetetur")
             .cx("et")
             .cr("sed")
             .c2coff("sit")
             .doit();

Methods

impl<'a, C, A> CseListCall<'a, C, A> where
C: BorrowMut<Client>,
A: GetToken,

pub fn doit(self) -> Result<(Response, Search)>

Perform the operation you have build so far.

pub fn q(self, new_value: &str) -> CseListCall<'a, C, A>

Query

Sets the q query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn start(self, new_value: u32) -> CseListCall<'a, C, A>

The index of the first result to return

Sets the start query property to the given value.

pub fn sort(self, new_value: &str) -> CseListCall<'a, C, A>

The sort expression to apply to the results

Sets the sort query property to the given value.

pub fn site_search_filter(self, new_value: &str) -> CseListCall<'a, C, A>

Controls whether to include or exclude results from the site named in the as_sitesearch parameter

Sets the site search filter query property to the given value.

pub fn site_search(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies all search results should be pages from a given site

Sets the site search query property to the given value.

pub fn search_type(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies the search type: image.

Sets the search type query property to the given value.

pub fn safe(self, new_value: &str) -> CseListCall<'a, C, A>

Search safety level

Sets the safe query property to the given value.

pub fn rights(self, new_value: &str) -> CseListCall<'a, C, A>

Filters based on licensing. Supported values include: cc_publicdomain, cc_attribute, cc_sharealike, cc_noncommercial, cc_nonderived and combinations of these.

Sets the rights query property to the given value.

pub fn related_site(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies that all search results should be pages that are related to the specified URL

Sets the related site query property to the given value.

pub fn or_terms(self, new_value: &str) -> CseListCall<'a, C, A>

Provides additional search terms to check for in a document, where each document in the search results must contain at least one of the additional search terms

Sets the or terms query property to the given value.

pub fn num(self, new_value: u32) -> CseListCall<'a, C, A>

Number of search results to return

Sets the num query property to the given value.

pub fn lr(self, new_value: &str) -> CseListCall<'a, C, A>

The language restriction for the search results

Sets the lr query property to the given value.

pub fn low_range(self, new_value: &str) -> CseListCall<'a, C, A>

Creates a range in form as_nlo value..as_nhi value and attempts to append it to query

Sets the low range query property to the given value.

pub fn link_site(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies that all search results should contain a link to a particular URL

Sets the link site query property to the given value.

pub fn img_type(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a type, which can be one of: clipart, face, lineart, news, and photo.

Sets the img type query property to the given value.

pub fn img_size(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a specified size, where size can be one of: icon, small, medium, large, xlarge, xxlarge, and huge.

Sets the img size query property to the given value.

pub fn img_dominant_color(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a specific dominant color: red, orange, yellow, green, teal, blue, purple, pink, white, gray, black and brown.

Sets the img dominant color query property to the given value.

pub fn img_color_type(self, new_value: &str) -> CseListCall<'a, C, A>

Returns black and white, grayscale, or color images: mono, gray, and color.

Sets the img color type query property to the given value.

pub fn hq(self, new_value: &str) -> CseListCall<'a, C, A>

Appends the extra query terms to the query.

Sets the hq query property to the given value.

pub fn hl(self, new_value: &str) -> CseListCall<'a, C, A>

Sets the user interface language.

Sets the hl query property to the given value.

pub fn high_range(self, new_value: &str) -> CseListCall<'a, C, A>

Creates a range in form as_nlo value..as_nhi value and attempts to append it to query

Sets the high range query property to the given value.

pub fn googlehost(self, new_value: &str) -> CseListCall<'a, C, A>

The local Google domain to use to perform the search.

Sets the googlehost query property to the given value.

pub fn gl(self, new_value: &str) -> CseListCall<'a, C, A>

Geolocation of end user.

Sets the gl query property to the given value.

pub fn filter(self, new_value: &str) -> CseListCall<'a, C, A>

Controls turning on or off the duplicate content filter.

Sets the filter query property to the given value.

pub fn file_type(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a specified type. Some of the allowed values are: bmp, gif, png, jpg, svg, pdf, ...

Sets the file type query property to the given value.

pub fn exclude_terms(self, new_value: &str) -> CseListCall<'a, C, A>

Identifies a word or phrase that should not appear in any documents in the search results

Sets the exclude terms query property to the given value.

pub fn exact_terms(self, new_value: &str) -> CseListCall<'a, C, A>

Identifies a phrase that all documents in the search results must contain

Sets the exact terms query property to the given value.

pub fn date_restrict(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies all search results are from a time period

Sets the date restrict query property to the given value.

pub fn cx(self, new_value: &str) -> CseListCall<'a, C, A>

The custom search engine ID to scope this search query

Sets the cx query property to the given value.

pub fn cr(self, new_value: &str) -> CseListCall<'a, C, A>

Country restrict(s).

Sets the cr query property to the given value.

pub fn c2coff(self, new_value: &str) -> CseListCall<'a, C, A>

Turns off the translation between zh-CN and zh-TW.

Sets the c2coff query property to the given value.

pub fn delegate(self, new_value: &'a mut dyn Delegate) -> CseListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> CseListCall<'a, C, A> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - Deprecated. Please use quotaUser instead.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.