Struct google_customsearch1::Customsearch

pub struct Customsearch<C, A> {
    // some fields omitted
}

Central instance to access all Customsearch related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_customsearch1 as customsearch1;
use customsearch1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use customsearch1::Customsearch;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Customsearch::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.cse().list("q")
             .start(26)
             .sort("erat")
             .site_search_filter("sadipscing")
             .site_search("dolor")
             .search_type("eirmod")
             .safe("elitr")
             .rights("amet")
             .related_site("no")
             .or_terms("labore")
             .num(62)
             .lr("dolore")
             .low_range("invidunt")
             .link_site("aliquyam")
             .img_type("accusam")
             .img_size("Lorem")
             .img_dominant_color("sea")
             .img_color_type("et")
             .hq("duo")
             .hl("et")
             .high_range("eirmod")
             .googlehost("sanctus")
             .gl("et")
             .filter("amet")
             .file_type("et")
             .exclude_terms("consetetur")
             .exact_terms("ut")
             .date_restrict("ea")
             .cx("sed")
             .cref("dolor")
             .cr("dolor")
             .c2coff("dolor")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Customsearch<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Customsearch<C, A>

fn cse(&'a self) -> CseMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/1.0.0.

Returns the previously set user-agent.