Struct google_customsearch1::CseListCall

pub struct CseListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Returns metadata about the search performed, metadata about the custom search engine used for the search, and the search results.

A builder for the list method supported by a cse resource. It is not used directly, but through a CseMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.cse().list("q")
             .start(5)
             .sort("amet.")
             .site_search_filter("voluptua.")
             .site_search("Lorem")
             .search_type("gubergren")
             .safe("justo")
             .rights("sit")
             .related_site("vero")
             .or_terms("diam")
             .num(35)
             .lr("consetetur")
             .low_range("sadipscing")
             .link_site("vero")
             .img_type("sadipscing")
             .img_size("invidunt")
             .img_dominant_color("consetetur")
             .img_color_type("dolore")
             .hq("duo")
             .hl("aliquyam")
             .high_range("Lorem")
             .googlehost("et")
             .gl("clita")
             .filter("consetetur")
             .file_type("takimata")
             .exclude_terms("nonumy")
             .exact_terms("kasd")
             .date_restrict("sanctus")
             .cx("takimata")
             .cref("At")
             .cr("labore")
             .c2coff("invidunt")
             .doit();

Methods

impl<'a, C, A> CseListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Search)>

Perform the operation you have build so far.

fn q(self, new_value: &str) -> CseListCall<'a, C, A>

Query

Sets the q query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn start(self, new_value: u32) -> CseListCall<'a, C, A>

The index of the first result to return

Sets the start query property to the given value.

fn sort(self, new_value: &str) -> CseListCall<'a, C, A>

The sort expression to apply to the results

Sets the sort query property to the given value.

fn site_search_filter(self, new_value: &str) -> CseListCall<'a, C, A>

Controls whether to include or exclude results from the site named in the as_sitesearch parameter

Sets the site search filter query property to the given value.

fn site_search(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies all search results should be pages from a given site

Sets the site search query property to the given value.

fn search_type(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies the search type: image.

Sets the search type query property to the given value.

fn safe(self, new_value: &str) -> CseListCall<'a, C, A>

Search safety level

Sets the safe query property to the given value.

fn rights(self, new_value: &str) -> CseListCall<'a, C, A>

Filters based on licensing. Supported values include: cc_publicdomain, cc_attribute, cc_sharealike, cc_noncommercial, cc_nonderived and combinations of these.

Sets the rights query property to the given value.

fn related_site(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies that all search results should be pages that are related to the specified URL

Sets the related site query property to the given value.

fn or_terms(self, new_value: &str) -> CseListCall<'a, C, A>

Provides additional search terms to check for in a document, where each document in the search results must contain at least one of the additional search terms

Sets the or terms query property to the given value.

fn num(self, new_value: u32) -> CseListCall<'a, C, A>

Number of search results to return

Sets the num query property to the given value.

fn lr(self, new_value: &str) -> CseListCall<'a, C, A>

The language restriction for the search results

Sets the lr query property to the given value.

fn low_range(self, new_value: &str) -> CseListCall<'a, C, A>

Creates a range in form as_nlo value..as_nhi value and attempts to append it to query

Sets the low range query property to the given value.

fn link_site(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies that all search results should contain a link to a particular URL

Sets the link site query property to the given value.

fn img_type(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a type, which can be one of: clipart, face, lineart, news, and photo.

Sets the img type query property to the given value.

fn img_size(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a specified size, where size can be one of: icon, small, medium, large, xlarge, xxlarge, and huge.

Sets the img size query property to the given value.

fn img_dominant_color(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a specific dominant color: yellow, green, teal, blue, purple, pink, white, gray, black and brown.

Sets the img dominant color query property to the given value.

fn img_color_type(self, new_value: &str) -> CseListCall<'a, C, A>

Returns black and white, grayscale, or color images: mono, gray, and color.

Sets the img color type query property to the given value.

fn hq(self, new_value: &str) -> CseListCall<'a, C, A>

Appends the extra query terms to the query.

Sets the hq query property to the given value.

fn hl(self, new_value: &str) -> CseListCall<'a, C, A>

Sets the user interface language.

Sets the hl query property to the given value.

fn high_range(self, new_value: &str) -> CseListCall<'a, C, A>

Creates a range in form as_nlo value..as_nhi value and attempts to append it to query

Sets the high range query property to the given value.

fn googlehost(self, new_value: &str) -> CseListCall<'a, C, A>

The local Google domain to use to perform the search.

Sets the googlehost query property to the given value.

fn gl(self, new_value: &str) -> CseListCall<'a, C, A>

Geolocation of end user.

Sets the gl query property to the given value.

fn filter(self, new_value: &str) -> CseListCall<'a, C, A>

Controls turning on or off the duplicate content filter.

Sets the filter query property to the given value.

fn file_type(self, new_value: &str) -> CseListCall<'a, C, A>

Returns images of a specified type. Some of the allowed values are: bmp, gif, png, jpg, svg, pdf, ...

Sets the file type query property to the given value.

fn exclude_terms(self, new_value: &str) -> CseListCall<'a, C, A>

Identifies a word or phrase that should not appear in any documents in the search results

Sets the exclude terms query property to the given value.

fn exact_terms(self, new_value: &str) -> CseListCall<'a, C, A>

Identifies a phrase that all documents in the search results must contain

Sets the exact terms query property to the given value.

fn date_restrict(self, new_value: &str) -> CseListCall<'a, C, A>

Specifies all search results are from a time period

Sets the date restrict query property to the given value.

fn cx(self, new_value: &str) -> CseListCall<'a, C, A>

The custom search engine ID to scope this search query

Sets the cx query property to the given value.

fn cref(self, new_value: &str) -> CseListCall<'a, C, A>

The URL of a linked custom search engine

Sets the cref query property to the given value.

fn cr(self, new_value: &str) -> CseListCall<'a, C, A>

Country restrict(s).

Sets the cr query property to the given value.

fn c2coff(self, new_value: &str) -> CseListCall<'a, C, A>

Turns off the translation between zh-CN and zh-TW.

Sets the c2coff query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> CseListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> CseListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.